fix: onboarding design and Avatar crash
- Fix Avatar crash: perform UInt modulo before Int conversion to prevent overflow
- Center-align all onboarding subtitle text to match design specification
- Make Skip button jump to card creation (slide 3) instead of exiting
- Fix s500 color definition (green channel 0.64 → 0.45 for proper slate gray)

+++ LinkMe/Components/Avatar.swift
@@ -16,7 +16,7 @@ struct Avatar: View {
     private static func toneFor(_ name: String) -> String {
         let tones = ["teal", "slate", "amber", "indigo", "rose", "sky"]
         let hash = name.utf8.reduce(0) { ($0 &* 31 &+ UInt($1)) }
-        return tones[Int(hash) % tones.count]
+        return tones[Int(hash % UInt(tones.count))]
     }
 
     private func toneColors() -> (bg: Color, fg: Color) {
